requirements.txt
anyjson
beautifulsoup4==4.7.1
coreapi==2.3.3
cython
django-chosen
django-crispy-forms
django-datatables-view
django-dirtyfields
django-extensions
django-filter
django-form-utils
django-gravatar2
django-guardian
django-hstore==1.3.5
django-notifications-hq
django-ratelimit==2.0.0
django-rest-swagger
django-rq
django-summernote==0.8.11.4
django-taggit
django-taggit-templatetags
django-user-agents
django>=1.11.18
djangorestframework
freegenes>=0.0.12
h5py
ipython
jsonschema
markdown
numexpr
oauth2client==3.0
Pillow
PyYAML==5.1
psycopg2-binary
pygments
python-jose==3.0.1
python3-saml
python-social-auth
requests
requests-oauthlib
requests-toolbelt
retrying
rq-scheduler
shapely
sendgrid==6.1.0
sentry-sdk==0.11.2
shippo==2.0.0rc1
social-auth-app-django
social-auth-core[saml]
geopy==1.20.0
uszipcode==0.2.2
uwsgi
